adding id checking and fixing repository visibility

+++ server/src/services/MongoService.ts
@@ -1,6 +1,5 @@
-import { Model } from "mongoose";
+import { Model, isValidObjectId } from "mongoose";
 import { IRepository } from "./types/IRepository";
-
 export class MongoService<T extends Model<any>> implements IRepository<T> {
   private readonly model
 
@@ -14,11 +13,15 @@ export class MongoService<T extends Model<any>> implements IRepository<T> {
   }
 
   async update(id: string, item: Partial<T>): Promise<Boolean> {
+    if(!isValidObjectId(id)) return false
+
     const updated = await this.model.updateOne({ _id: id }, item)
     return updated.modifiedCount >= 1
   }
 
   async delete(id: string): Promise<Boolean> {
+    if(!isValidObjectId(id)) return false
+
     const deleted = await this.model.deleteOne({ _id: id }).exec()
     return deleted.deletedCount === 1
   }
@@ -29,6 +32,8 @@ export class MongoService<T extends Model<any>> implements IRepository<T> {
   }
 
   async findOne(id: string): Promise<T | null> {
+    if(!isValidObjectId(id)) return null
+
     const document = await this.model.findById(id)
     return document
